backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-docx==1.0.1
openpyxl==3.1.2
pandas==2.1.1
pillow==10.1.0
python-multipart==0.0.6
pydantic==2.4.2